requirements.txt
# Required
openai-whisper
pyinstaller
# Optional
ffmpeg-python
setuptools-rust